# specs/002-enterprise-uiux-overhaul/contracts/gateway-uiux.openapi.yaml
openapi: 3.0.3
info:
  title: Orbit Tasks Gateway v2 Contract
  version: 2.0.0
  description: Public REST contract for Sprint-DSU loop, work item multiview CRUD, collaboration triage, and schedule intelligence.
servers:
  - url: https://tasksapi.infinitefallcult.trade
    description: Production gateway
  - url: http://localhost:8080
    description: Local gateway
paths:
...
  /api/v2/threads:
    post:
      summary: Create thread
      operationId: createThreadV2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId, workItemId, title, createdBy]
              properties:
                workspaceId:
                  type: string
                  format: uuid
                workItemId:
                  type: string
                  format: uuid
                title:
                  type: string
                createdBy:
                  type: string
      responses:
        "200":
          description: Thread
  /api/v2/threads/{threadId}/messages:
    get:
      summary: List thread messages
      operationId: listThreadMessagesV2
      parameters:
        - in: path
          name: threadId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Messages
    post:
      summary: Post thread message
      operationId: postThreadMessageV2
      parameters:
        - in: path
          name: threadId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [authorId, body]
              properties:
                authorId:
                  type: string
                body:
                  type: string
      responses:
        "200":
          description: Message
  /api/v2/inbox:
    get:
      summary: List inbox items by filter
      operationId: listInboxV2
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: filter
          schema:
            type: string
            enum: [all, notifications, requests, mentions, ai_questions]
      responses:
        "200":
          description: Inbox items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InboxItem"
  /api/v2/inbox/{inboxItemId}:
    patch:
      summary: Patch inbox status
      operationId: patchInboxItemV2
      parameters:
        - in: path
          name: inboxItemId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
                status:
                  type: string
                note:
                  type: string
      responses:
        "200":
          description: Updated inbox item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItem"
...
components:
...
  schemas:
...
    InboxItem:
      type: object
      properties:
        inboxItemId:
          type: string
        userId:
          type: string
        kind:
          type: string
        sourceType:
          type: string
        sourceId:
          type: string
        messageId:
          type: string
        read:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
        resolvedAt:
          type: string
          nullable: true
...